Attempt to handle errors
This is still kludgy.  If SHELL cannot be exec'd, the error message will
go to the pty that will be destroyed before the user can see it.  Perhaps
we should set no_prune on the root, but we shouldn't do that until we
figure out how we want to subsequently prune the window.  This is probably
not really worth thinking about, and I'm just bike shedding while procrastinating.