Stop reshaping windows
I think it makes more sense to let the underlying pty keep its original shape
when the window resizes.  We should probably reshape if the window grows
larger than the pty, but punt on that for now.  Also, this makes the
behavior of a split somewhat more natural in my opinion, and that
change is reflected in the test suite change seen here.