Reshape_window to make height match
See the change in test suite.  This is better behavior.  If a pager is running,
we want it to take a SIGWINCH and know the height of the window.  Also, if we
are going to run vi or similar, the pty will always need to know the height
accurately.  We can make an action required to modify the width, which will
be in the next commit.