Do not return a value from actions
At the moment, we are completely ignoring these.  I think the right
thing to do is simply print an error message if an error occurs in
an action.  Perhaps, when actions are more involved, it will make
senese to return a value.  If/when that is the case, bring back
return values