Set title after changing offset
The problem (which persists) is that a terminal can change the
focus point and trigger an offset change which our title does not
reflect.  eg, man bash will open the alternate screen and position
the cursor not in the first column.  If the position is beyond the
extent of the window and manual scroll is off, the user gets a
window in which the cursor is visible but the title still shows
the display as column 1.  This patch fixes that slightly, but
the title are still getting munged.  Probably need to redraw
the title always. :(