Read more than one byte at a time
This change makes it easier to examine the context of a match
in a debugger, and I can now explain why we match 3 times.

When matching, the context of the matches looks like:

yes | nl -ba | sed 400q
^G21
echo 123456789
yes | nl -ba | sed 400q^Mecho 123456789
   380381382383384385386387388389390391392393394395396397398399400^M/bin/sh$
   123456789

The first match is the pty echoing back what is typed, the 2nd
is smtx displaying the text, and the 3rd is the shell's output.
(Hmm.  This is not satisfying either.  "smtx displaying the text"
sounds like gobbledygook)