Pass smtx as TERM for test_nel
I am fairly sure that ##__VA_ARGS__ is a gnu extension for gcc, but
at the moment do not know a cleaner way to do this.  Although it might
be cleaner to drop the macro and just call new_test directly. (I just
prefer using the stringization operator ensure the test is named
accurately.  OTOH, all of this cruft feels kludgy and ought to be
redone.  Perhaps just rip out the logic allowing tests to be specified and just
run them all.