Use err_check() in rewrite()
This is probably silly.  The primary purpose of this mostly
trivial refactor is to use err_check so that we avoid having
to figure out a way to get test coverage of the failure condition.
However, in the current test suite, this loop is typically executed
exactly once and none of this logic is genuinely tested.