Add copyright to action.c
It is not clear to me that I should include Rob King in this copyright,
since (I believe) all of the code in that file is actually authored by
me.  But I am not a not a lawyer, and am erring on the side of caution.

To reiterate the history of this project, smtx is derived from mtm, and
it seems reasonable to extend Rob King's copyright to the entire code
base.