Remove invalid terminfo entry
The syntax is wrong.  It should be "U8@, use=smtx", rather than
"@U8, use=smtx", so I suspect this is just wrong.  Currently
getting errors with tic that U8 and E3 are unknown capabilities.
I do not claim to be an expert in this, so for now I will just
delete the entries.