Reorder state assignments for escape
There are some non-trivial changes here.  In particular, we had previously
overwritten the look up table entry for 0x21.  This was because of a
misunderstanding from the orginal code in mtm.  In the MAKESTATE macro
from cae64949c9d5801faaef6b1c606a01049bf7eeea, the states are searched
linearly for a match, so individual assignments in the macro that
occur earlier supersede later entries, but in the lut construction
they incorrectly get overwritten.  We will probably need to go though
this carefully at some point to ensure that no changes were made in
the migration.