Use osc string to set titles
This is a terrible implementation, but allows us to set titles via
osc sequences.  eg printf '\033]2foobar\007'

Don't have a lot of time right now to clean this up, so consider this
merely proof of concept.  Not adding any tests since I am certain this
is not what we actually want to do.