Return size_t from the desciption functions
This is a response to the compiler warning about the return
value being uninitialized in describe_row.  It seems appropriate,
since this value is passed as the 3rd argument to write, which expects
a size_t.  We are not using -1 as an error value, so there is no point
in returning it.