Set TERM=smtx if available and use newterm()
By default, we should use the smtx terminfo if they are available.  If
not, fall back to the settings of the current terminal.  I suppose
a long term goal is to make this usable as a login shell, but we
can cross that bridge when we come to it.  This is good for now.