Remove mode from struct state
We want to incapsulate the mode in the key bindings.  The "sink" mode
was speculative and unused, and I think undesirable.  If we want to do that,
we should assign a noop action to the keys in the binding.  Indeed, we
should probably assign an action to every key in all bindings and skip
the else portion of the main loop.