Store errmsg in state
Bike shedding.  It feels cleaner to keep the error message this way.  Since
we are now using S.werr to display commands, we should rename it and
change the logic.  (It is not clean to have the command mode key bindings
calling err_check to copy the command into the error message!)