Get endwin out of atexit handler
Now that the test suite is not sending signals, we can skip this cruft.  Maybe it
is cleaner to stay in the atexit handler, but right now I want to get a good
error message on memory allocation error for the main window, and this seems to
be getting in the way.