Use more distinct prompts in test_lnm()
This is frustrating.  I saw a test run on termux in which the received text
in row 8 was "u2>syn".  I cannot explain that except to speculate that the
"u2>" text was similar to "u1>" and ncurses took some optimizations which
caused the string "u2>" to appear early and allow the test to proceed
before the shell had actually printed the prompt.  Either that, or the
shell is printing the prompt twice.