Stub in osc calls for show_layout, etc.
This currently cannot work because the byte sequences written to the master pty are
interpreted as keystrokes.  Currently, they will only be vtwritten if they are
read out of a shell because of the special case handling in getinput().
It is not clear to me how best to resolve this special case (which
should not be special).