Use more unique strings
I was getting a log of timeouts.  It seems that ncurses is optimizing
outputs and perhaps reusing the existing content.  By not re-using "un",
we get fewer timeouts.  However, since it appears that all of the output
to the pty is somewhat non-deterministic, all of these tests are inherently
fragile.