In grep, push in on first match miss
The summary of this commit is a stellar example of confuscated English.  Consider
the case where we are trying to match the needle "abc" and the input string
is "aabc".  When we see the first "a", the second "a" fails to match
the "b", so we reset the needle.  But it does match the first char, so
we need to be looking in the right place.  This grep is not intended
to be robust, but we at least need to to catch this simple case.