Replace goto (pnl) with fallthrough
Hmmm.  The 3 options seem to be invoking CALL, gotos, or this fallthrough.  I am
torn on which is worse.  Perhaps we should bite the bullet and split nel and
ind into functions.  I suspect that long term we will wind up with having
a complete dispatch table and this entire giant switch will be discarded.  For
now, let's keep it like this.