Fix handling of incomplete character
Well, this does not really fix the handling of it, but it more accurately
represents what is actually happening.  The old comment (inherited from mtm)
stated "try again", but the string was never actually retried.  Instead, vtwrite
returned and the data was discarded.  Stepping over the first byte of the
incomplete character seems pointless, but somehow writing VTPARSER_BAD_CHAR
to the tty seems more appropriate than just discarding all the data.