Avoid unnecessary assignment
This is mostly for code clarity.  The assignment to zero is redundant
because it was set in reset(), and the relevant portion of this
code is the side effect.  This code seems overly convoluted, since
it is trying to increment nargs either on a semi-colon or for the first
value.