Add a binding to show_state
If we wish to remove command mode, we will not be able to call show_layout
to get the name of the slave side of the pty, so printing it in the layout
will not work.  Instead, we could put it in the state info and query it
with a key binding. We probably want a better mechanism for getting status.