Do not require root proc to be alive
Prior to this commit, the main program would exit if we close the
shell in the pty with the highest fd (eg, if S.p->fd gets set to -1).
Before 35091738f044c54ac07bdf2b5d326fdf3e0224e8, we avoided this
by pushing the dead pty to the end of the list so that we would
only exit if all pty are dead. However, if we want to retain error
messages, it seems appropriate to not quit even if all shells have
terminated.

However, now when we close the last shell the program does not
exit but it hangs waiting for input.  The test suite still passes. :(
We need to add a test for this and fix it.