Stop sorting the proclist: keep maxfd instead
This simplifies the code substantially.  We are now
not precise with maxfd, as we never decrease it, but that
is a trivial performance hit in select.  This commit
also makes it so that the program does not exit when
all shells are done.  I am undecided about what the
behavior should be, but I think this might be desired.