Modify test_quit to reflect new usage
This test is a bit confusing.   We close the shell, and then must q to cause
an exit since the test driver fall back will not be able to use the shell
that has now exited.  This is desired behavior.  If we change the code
so that the q will not cause smtx to exit, then this test will fail with
a timeout.