Adjust balance()
When the last canvas in the chain has an unexpected typ, we
ought not abort the balance early.  Note that we really
need to clean this up, as I am unsure what the right behavior
is.  But this is certainly better.  Maybe we need typ to
be a tri-state item with "undetermined" as an option rather
than defaulting to 0 for canvas with no children.