Keep pty ordered on proc list
This is not quite right.  We probably want to make the proc
list immutable and make the free list redundant. (By which I mean
that items are on the free list and the proc list simultaneously.)
Or....dump the free list completely and just walk the list looking
at counts when trying to get an available pty.)