Always navigate by tree
The logic for searching by position is overly complex, and navigating
by tree makes the tree structure more obvious to the user.  It will
take a little getting used to, but it is good to train the user's brain
to recognize the tree structure.  Also, we could to implement a "goto"
command to ease navigation, if desired.