Stop using newterm()
We want an early failure if the terminal type does not exist.  I believe
this code was introduced so that the test suite would work if the
smtx terminfo was not installed, but it causes undesirable behavior and
is overly complicated.  If the user has an incorrect TERM setting,
the program should fail.