Stub in passing row count to newcanvas
Something strange is afoot.  In mtm, the alternate screen
is initialized with a window only as large as the current
canvas.  We do not know the size when new_pty is invoked,
so have been punting and overallocating.  When we try to
set p->ws.ws_row in new_pty, the tests fail in obscure
ways.

We need to figure this out, and simplest solution seems to be
to know the size of the canvas when we create the pty.  That,
or dump scroll bottom and keep track of position more accurately.

It is not really clear yet what is the proper way forward.