Put initial window at top of screen
It makes everything easier if we do not try to do all of the
calculation regarding top of screen.  This will allow a lot
of cruft to be cleared out of handler.c, and should make
new windows line up better.  (eg, the initial prompt will
always be visible instead of stuck in the history off the
top of the display)