Use for loop to get block in switch
This is a stylistically challenging point.  I hate using a block inside
the switch to declare a variable, so this is abusing the language to
get the block by using a for loop. Not really sure this is a good
idea, but ...  it lets us get rid of the extra braces!