Hide unimplemented members of cmd enum
We are (tentatively?) removing the implementation of these functions, since
they seem unnecessary.  Will (for the moment) leave them in the enum so we
do not completely forget their existence.  However, we get a compiler
warning about the cases not be handled in the switch, so we hide them
behind the preprocessor directive to avoid that.  Should probably just
delete them.