Call endwin in atexit handler, exit on SIGTERM
Not really sure that we should be catching SIGTERM at all.  Perhaps
the better behavior is to just let it die and not clean up the
windows.  That would allow the unexpecting user to see what was
left on screen at least.  A SIGTERM should be considered an exceptional
condition.