Use switch on first char to transition
Seems really ugly to do a string of strcmps.  Not worth building
a full hash table.

Note that we changed the test because things start to timeout because
we removed the refresh.  I think that not refreshing is the correct
behavior.  We should had a command to explicitly refresh rather
than doing it on all transitions.