Defer creation of pty to after canvas
This change is the first step towards getting new_pty() out of newcanvas().
Eventually, we want to be able to use the extent of the canvas to create
the pty with the correct size on initialization, but we will need to work
out the details in the balance().  Baby steps.