Fix argv0 in test
We cannot set argv0 to the local buffer.  At this point, it is not
clear to me how this test suite was ever passing on any machine.  I
suspect I was too lax when refactoring.  I am fairly confident
that the test suite was working in the debian docker image on Macos
a few months ago, but docker is borked at the moment and I cannot
verify that.  Test suite is failing on centos.